+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+### Deprecations
+
+- [Schema] Column type 'bool' is deprecated — change to 'boolean'
+
 ## Unreleased
 
 ### Deprecations